+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- For nexus, only use pickup points tagged 'TaxJar'
+
 ## [0.0.6] - 2021-05-19
 
 ### Changed